Not all elements in a folder are subfolders..

+++ arrowverse_reorderer/__main__.py
@@ -104,6 +104,8 @@ def reorder(folders, name_dest_folder, destination_path, dry_run=False):
 
 		# if (sub)folder is empty, delete it
 		for subfolder in os.listdir(folder):
+			if not os.isdir(subfolder):
+				continue
 			if len(os.listdir(subfolder)) == 0:
 				os.rmdir(subfolder)
 				print("Folder "+subfolder+" deleted\n")
